backend/requirements.txt
asgiref==3.11.1
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cryptography==46.0.5
dj-database-url==3.1.2
Django==5.2.12
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework-csv==3.0.2
drf-spectacular==0.29.0
drf-spectacular-sidecar==2026.3.1
et_xmlfile==2.0.0
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
numpy==2.4.2
openpyxl==3.1.5
pandas==3.0.1
pycparser==3.0
PyJWT==2.11.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
unicodecsv==0.14.1
uritemplate==4.2.0
urllib3==2.6.3
whitenoise==6.8.2
